[IMP] fleet_vehicle_service_kanban: Better organization of tests code.
