+++ fleet_vehicle_service_kanban/tests/__init__.py
@@ -1,4 +1,5 @@
 # Copyright 2020-Present Druidoo - Manuel Marquez <manuel.marquez@druidoo.io>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
+from . import common
 from . import test_vehicle_log_services
